requirements.txt
discord.py>=2.4,<3
python-dotenv>=1.0,<2
Pillow>=10,<12
